refactor(test): 메인 패키지 구조에 맞게 테스트 구조 변경
