+++ src/test/java/lotto/util/ValidatorTest.java
@@ -1,9 +1,6 @@
-package lotto;
+package lotto.util;
 
-import lotto.util.Validator;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.*;
